move "tryCollectImportsDeep" to core
