+++ packages/build-tools/src/index.ts
@@ -2,5 +2,4 @@ export { processUrlDependencies } from './process-url-dependencies';
 export { hasImportedSideEffects } from './has-imported-side-effects';
 export { sortModulesByDepth } from './sort-modules-by-depth';
 export { loadStylableConfig } from './load-stylable-config';
-export { tryCollectImportsDeep } from './try-collect-imports-deep';
 export { CalcDepthContext, calcDepth, getCSSViewModule } from './calc-depth';
