--- packages/build-tools/src/try-collect-imports-deep.ts
@@ -1,18 +1,0 @@
-import type { Stylable, StylableMeta } from '@stylable/core';
-
-export function tryCollectImportsDeep(
-    stylable: Stylable,
-    meta: StylableMeta,
-    imports = new Set<string>()
-) {
-    for (const { context, request } of meta.getImportStatements()) {
-        try {
-            const resolved = stylable.resolver.resolvePath(context, request);
-            imports.add(resolved);
-            tryCollectImportsDeep(stylable, stylable.analyze(resolved), imports);
-        } catch (e) {
-            /** */
-        }
-    }
-    return imports;
-}
