+++ packages/cli/src/build.ts
@@ -1,3 +1,4 @@
+import { tryCollectImportsDeep } from '@stylable/core/dist/index-internal';
 import type { BuildContext, BuildOptions } from './types';
 import { IndexGenerator as BaseIndexGenerator } from './base-generator';
 import { generateManifest } from './generate-manifest';
@@ -8,7 +9,6 @@ import { DiagnosticsManager } from './diagnostics-manager';
 import type { Diagnostic } from './report-diagnostics';
 import { tryRun } from './build-tools';
 import { errorMessages, buildMessages } from './messages';
-import { tryCollectImportsDeep } from '@stylable/build-tools';
 
 export async function build(
     {
