+++ packages/core/src/index-internal.ts
@@ -2,4 +2,5 @@ export { safeParse } from './parser';
 export { getAstNodeAt } from './helpers/ast';
 export { process } from './stylable-processor';
 export { StylableTransformer } from './stylable-transformer';
+export { tryCollectImportsDeep } from './helpers/import';
 export type { MappedStates } from './features';
