+++ packages/rollup-plugin/src/index.ts
@@ -2,12 +2,12 @@ import type { Plugin } from 'rollup';
 import fs from 'fs';
 import { join, parse } from 'path';
 import { Stylable, emitDiagnostics, DiagnosticsMode } from '@stylable/core';
+import { tryCollectImportsDeep } from '@stylable/core/dist/index-internal';
 import {
     sortModulesByDepth,
     calcDepth,
     CalcDepthContext,
     hasImportedSideEffects,
-    tryCollectImportsDeep,
 } from '@stylable/build-tools';
 import { resolveNamespace as resolveNamespaceNode } from '@stylable/node';
 import { StylableOptimizer } from '@stylable/optimizer';
